Fixed up tests for refactored exceptions
